Update LimitDefault and LimitMax type from uint32 to uint64 (#12307)
* Updated LimitDefault and LimitMax type from uint32 to uint64

* Fixed linting

* Fixed integration tests

* Cleaned unnecessary uint64 conversions

* Fixed mock

* Fixed linting

* Fixed config test

* Fixed gas limit multiplier method

* Fixed linting and tests

* Fixed confirmer tests

* Fixed VRF tests

* Fixed linting

* Undid unnecessary method signature changes

* Fixed linting

+++ core/services/relay/evm/evm.go
@@ -498,10 +498,10 @@ func newOnChainContractTransmitter(ctx context.Context, lggr logger.Logger, rarg
 	gasLimit := configWatcher.chain.Config().EVM().GasEstimator().LimitDefault()
 	ocr2Limit := configWatcher.chain.Config().EVM().GasEstimator().LimitJobType().OCR2()
 	if ocr2Limit != nil {
-		gasLimit = *ocr2Limit
+		gasLimit = uint64(*ocr2Limit)
 	}
 	if opts.pluginGasLimit != nil {
-		gasLimit = *opts.pluginGasLimit
+		gasLimit = uint64(*opts.pluginGasLimit)
 	}
 
 	transmitter, err := ocrcommon.NewTransmitter(
